Friender-Backend/app.py
import os
from dotenv import load_dotenv

from flask import (
    Flask, request, redirect, session, g
)
# from flask_debugtoolbar import DebugToolbarExtension

# from forms import (
#     UserAddForm, UserEditForm, LoginForm, MessageForm, CSRFProtection,
# )
from models import (
    connect_db
)

load_dotenv()

app = Flask(__name__)

app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv("DB_URL")
app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
app.config['SQLALCHEMY_ECHO'] = True

connect_db(app)

##############################################################################

##############################################################################

##############################################################################

##############################################################################

@app.get('/')
def test():
  return "hello"